+++ CHANGELOG.md
@@ -4,9 +4,13 @@ All notable changes to this project will be documented in this file. Items under
 Contributors: please follow the recommendations outlined at [keepachangelog.com](http://keepachangelog.com/). Please use the existing headings and styling as a guide, and add a link for the version diff at the bottom of the file. Also, please update the `Unreleased` link to compare to the latest release version.
 ## [Unreleased]
 
+## [1.1.0] - 2016-08-01
+##### Added
+- Support for custom .bootstraprc location. [#114](https://github.com/shakacode/bootstrap-loader/pull/114) by [justin808](https://github.com/justin808) and [pherris](https://github.com/pherris).
+
 ## [1.0.10] - 2016-03-17
 ##### Fixed
-- fixes polyfill require for node < v4.0.0. [#72](https://github.com/shakacode/bootstrap-loader/pull/72) by [mdgraser](https://github.com/mdgraser).
+- Fixes polyfill require for node < v4.0.0. [#72](https://github.com/shakacode/bootstrap-loader/pull/72) by [mdgraser](https://github.com/mdgraser).
 
 ## [1.0.9] - 2016-02-28
 ##### Fixed
@@ -14,7 +18,8 @@ Contributors: please follow the recommendations outlined at [keepachangelog.com]
 
 ## [1.0.8]
 
-[Unreleased]: https://github.com/shakacode/bootstrap-loader/compare/1.0.10...master
+[Unreleased]: https://github.com/shakacode/bootstrap-loader/compare/1.1.0...master
+[1.1.0]: https://github.com/shakacode/bootstrap-loader/compare/1.0.10...1.1.0
 [1.0.10]: https://github.com/shakacode/bootstrap-loader/compare/1.0.9...1.0.10
 [1.0.9]: https://github.com/shakacode/bootstrap-loader/compare/1.0.8...1.0.9
 [1.0.8]: https://github.com/shakacode/bootstrap-loader/compare/1.0.7...1.0.8